refactor: 예외 처리 로직 위치 변경
